Move most of the logic out of __main__.py, and into builder.py
This is needed to that I can import the builder into potential unit tests.
